feat: gate egui under egui feature flag
