+++ Cargo.toml
@@ -16,11 +16,6 @@ categories = ["multimedia::images"]
 [dependencies]
 anyhow = "1.0.75"
 async-trait = "0.1.74"
-eframe = "0.29.1"
-egui-theme-switcher = "0.1.0"
-egui-file-dialog = "0.7.0"
-egui_extras = {version = "0.29.1", features = ["default", "all_loaders"] }
-image = { version = "0.25.1", features = ["jpeg", "png", "gif"] } 
 rand = "0.8.5"
 reqwest = "0.12.4"
 rss = "2.0.6"
@@ -30,6 +25,18 @@ strum_macros = "0.26.2"
 thiserror = "2.0.3"
 tokio = { version = "1.33.0", features = ["fs", "macros", "rt-multi-thread"]}
 tokio-util = "0.7.11"
+
+#egui deps
+eframe = {version = "0.29.1", optional = true}
+egui-theme-switcher = { version = "0.1.0", optional = true}
+egui-file-dialog = { version = "0.7.0", optional = true}
+egui_extras = {version = "0.29.1", features = ["default", "all_loaders"], optional = true }
+image = { version = "0.25.1", features = ["jpeg", "png", "gif"], optional = true} 
+
+
+[features]
+default=["egui"]
+egui=["eframe", "egui-theme-switcher", "egui-file-dialog", "egui_extras", "image"]
 
 # Config for 'cargo dist'
 [workspace.metadata.dist]
